refactor: simplify HilingualBannerAd and move logic to BannerAdState
- Relocate `createAndLoadAdView`, `createAdLoadCallback`, and `getAdSize` from `HilingualBannerAd.kt` to `BannerAdState.kt`
- Overload `HilingualBannerAd` to support both automatic `BannerAdHolder` creation and direct injection
- Remove manual ad loading and resource release logic from `HilingualBannerAd` in favor of state-managed lifecycle
- Update `GrammarSpellingTab` and `RecommendExpressionTab` to use the simplified `HilingualBannerAd` component with nullable `adHolder` checks
