+++ core/ui/src/main/java/com/hilingual/core/ui/component/item/diary/tab/RecommendExpressionTab.kt
@@ -96,10 +96,7 @@ fun RecommendExpressionTab(
         if (isAdVisible) {
             item {
                 Spacer(Modifier.height(24.dp))
-                HilingualBannerAd(
-                    adUnitId = BuildConfig.ADMOB_INLINEBANNER_UNIT_ID,
-                    adHolder = adHolder,
-                )
+                if (adHolder != null) HilingualBannerAd(adHolder = adHolder)
             }
         }
     }
